ai_service/requirements.txt
# ============================================================================
# FILE: ai_service/requirements.txt
# Lists the Python packages required for this service.
# ============================================================================
"""
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
requests
"""